requirements.txt
PyPDF2
python-docx
scikit-learn==1.2.2
openai==0.27.8
python-dotenv==1.0.0
anthropic==0.43.1
numpy==1.24.0
pandas==1.5.3
gensim==4.3.1
spacy
nltk
streamlit
